Added code to create logs for each MPI rank. Logs come from a fake particle initialization with each ranks initialized with different values. Log files consist of a header containing titles of objects and the data separated by space.
Originally created with the intent of showing how one could
achieve coupling between different entities - a component that
appends a _rank at the end of the log files, and another that
prints out the log files - albeit loosely, with namespaces, and not
by putting them inside a class. This is an idea that motivates itself
by recognizing that there is a relationship between the data, but not
quite one that one would associate between classes.

While creating the code, I made rudimentary use of other features such
as 'map'. I hope to create some use cases for STL in the future.

The namespace idea was borrowed from an item in Scott Meyers' 'Effective C++' book,
where he makes the very sensible case of applying groupings based on
common sense, and not viewing them as end in themselves.